+++ momento-sdk/src/main/java/momento/sdk/auth/ApiKeyV2CredentialProvider.java
@@ -1,10 +1,12 @@
 package momento.sdk.auth;
 
-import java.util.Base64;
+import static momento.sdk.internal.AuthUtils.isV2ApiKey;
+
 import javax.annotation.Nonnull;
 import momento.sdk.exceptions.InvalidArgumentException;
 
 public class ApiKeyV2CredentialProvider extends CredentialProvider {
+
   private final String authToken;
   private final String controlEndpoint;
   private final String cacheEndpoint;
@@ -15,76 +17,17 @@ public class ApiKeyV2CredentialProvider extends CredentialProvider {
     return prefix + "." + endpoint;
   }
 
-  private static boolean isV2ApiKey(String authToken) {
-    try {
-      // only v1 api keys are entirely b64 encoded
-      // v2 keys are JWTs with b64 encoded segments
-      if (isBase64Encoded(authToken)) {
-        return false;
-      }
-
-      // JWT tokens have 3 parts separated by dots
-      if (authToken.chars().filter(ch -> ch == '.').count() != 2) {
-        return false;
-      }
-
-      // Split and get the payload (second part)
-      String[] parts = authToken.split("\\.");
-      if (parts.length != 3) {
-        return false;
-      }
-
-      // Decode the payload from base64
-      String payload = new String(Base64.getUrlDecoder().decode(parts[1]));
-
-      // Check if it contains "t":"g" (global key indicator)
-      return payload.contains("\"t\"") && payload.contains("\"g\"");
-    } catch (Exception e) {
-      return false;
-    }
-  }
-
-  public static boolean isBase64Encoded(String apiKey) {
-    try {
-      Base64.getUrlDecoder().decode(apiKey);
-      return true;
-    } catch (IllegalArgumentException e) {
-      return false;
-    }
-  }
-
-  // private static boolean isBase64EncodedToken(String apiKey) {
-  // // Check if it's a global JWT (which is allowed)
-  // if (isV2ApiKey(apiKey)) {
-  // return false;
-  // }
-
-  // // Legacy tokens have format: xxx.yyy.zzz (JWT format)
-  // if (apiKey.chars().filter(ch -> ch == '.').count() == 2) {
-  // return true;
-  // }
-
-  // // Check if it's base64 encoded (V1 tokens are base64 encoded)
-  // try {
-  // Base64.getUrlDecoder().decode(apiKey);
-  // return true;
-  // } catch (Exception e) {
-  // return false;
-  // }
-  // }
-
   public ApiKeyV2CredentialProvider(@Nonnull String authToken, @Nonnull String endpoint) {
-    if (authToken == null || authToken == "") {
-      throw new InvalidArgumentException("Auth token must not be empty");
+    if (authToken.isEmpty()) {
+      throw new InvalidArgumentException("API key must not be empty");
     }
-    if (endpoint == null || endpoint == "") {
+    if (endpoint.isEmpty()) {
       throw new InvalidArgumentException("Endpoint must not be empty");
     }
 
     if (!isV2ApiKey(authToken)) {
       throw new InvalidArgumentException(
-          "V2 API key appears to be a V1 or legacy token. "
-              + "Did you mean to use CredentialProvider.fromString() or CredentialProvider.fromEnvVar() instead?");
+          "Received an invalid v2 API key. Did you mean to use `fromString()` or `fromEnvVar()` with a legacy key instead?");
     }
 
     this.authToken = authToken;
