+++ momento-sdk/src/main/java/momento/sdk/auth/EnvVarV2CredentialProvider.java
@@ -5,18 +5,18 @@ import momento.sdk.exceptions.InvalidArgumentException;
 
 public class EnvVarV2CredentialProvider extends ApiKeyV2CredentialProvider {
 
-  public EnvVarV2CredentialProvider(@Nonnull String envVarName, @Nonnull String endpointEnvVar) {
-    super(getApiKeyValueFromEnvVar(envVarName), getEndpointValueFromEnvVar(endpointEnvVar));
+  public EnvVarV2CredentialProvider(@Nonnull String apiKeyEnvVar, @Nonnull String endpointEnvVar) {
+    super(getApiKeyValueFromEnvVar(apiKeyEnvVar), getEndpointValueFromEnvVar(endpointEnvVar));
   }
 
-  private static String getApiKeyValueFromEnvVar(String envVarName) {
-    if (envVarName == null || envVarName.isEmpty()) { // Check for empty env var name
+  private static String getApiKeyValueFromEnvVar(String apiKeyEnvVar) {
+    if (apiKeyEnvVar == null || apiKeyEnvVar.isEmpty()) { // Check for empty env var name
       throw new InvalidArgumentException("Env var name cannot be empty");
     }
 
-    String authToken = System.getenv(envVarName);
+    String authToken = System.getenv(apiKeyEnvVar);
     if (authToken == null || authToken.isEmpty()) { // Check for empty value
-      throw new InvalidArgumentException("Env var " + envVarName + " must be set");
+      throw new InvalidArgumentException("Env var " + apiKeyEnvVar + " must be set");
     }
     return authToken;
   }
